updated partly UI
