+++ package.json
@@ -1,44 +1,44 @@
 {
-  "name": "bee2bee",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "@emotion/react": "11.11.0",
-    "@emotion/styled": "11.11.0",
-    "@fireworks-js/react": "^2.10.7",
-    "@mui/icons-material": "5.11.16",
-    "@mui/material": "5.13.3",
-    "@mui/styled-engine-sc": "5.12.0",
-    "@next/bundle-analyzer": "^13.5.6",
-    "@reduxjs/toolkit": "^1.9.6",
-    "axios": "1.4.0",
-    "browser-image-compression": "^2.0.2",
-    "dotenv": "16.3.1",
-    "eslint": "8.41.0",
-    "eslint-config-next": "13.4.4",
-    "js-cookie": "^3.0.5",
-    "moment": "^2.29.4",
-    "moment-timezone": "^0.5.43",
-    "next": "13.4.4",
-    "react": "18.2.0",
-    "react-confetti": "^6.1.0",
-    "react-dom": "18.2.0",
-    "react-google-recaptcha": "^3.1.0",
-    "react-markdown": "8.0.7",
-    "react-redux": "^8.1.3",
-    "react-snowfall": "^1.2.1",
-    "redux": "^4.2.1",
-    "sharp": "^0.32.6",
-    "styled-components": "5.3.11",
-    "swr": "^2.2.4"
-  },
-  "devDependencies": {
-    "webpack-bundle-analyzer": "^4.9.1"
-  }
+	"name": "bee2bee",
+	"version": "0.1.0",
+	"private": true,
+	"scripts": {
+		"dev": "next dev",
+		"build": "next build",
+		"start": "next start",
+		"lint": "next lint"
+	},
+	"dependencies": {
+		"@emotion/react": "11.11.0",
+		"@emotion/styled": "11.11.0",
+		"@fireworks-js/react": "^2.10.7",
+		"@mui/icons-material": "5.11.16",
+		"@mui/material": "5.13.3",
+		"@mui/styled-engine-sc": "5.12.0",
+		"@next/bundle-analyzer": "^13.5.6",
+		"@reduxjs/toolkit": "^1.9.6",
+		"axios": "1.4.0",
+		"browser-image-compression": "^2.0.2",
+		"dotenv": "16.3.1",
+		"js-cookie": "^3.0.5",
+		"moment": "^2.29.4",
+		"moment-timezone": "^0.5.43",
+		"next": "13.4.4",
+		"react": "18.2.0",
+		"react-confetti": "^6.1.0",
+		"react-dom": "18.2.0",
+		"react-google-recaptcha": "^3.1.0",
+		"react-markdown": "8.0.7",
+		"react-redux": "^8.1.3",
+		"react-snowfall": "^1.2.1",
+		"redux": "^4.2.1",
+		"sharp": "^0.32.6",
+		"styled-components": "5.3.11",
+		"swr": "^2.2.4"
+	},
+	"devDependencies": {
+		"webpack-bundle-analyzer": "^4.9.1",
+		"eslint": "8.41.0",
+		"eslint-config-next": "13.4.4"
+	}
 }
